Fix camera absolute position
If timmeeSinceLastUpdate is initialized with performance.now the camera
position will be in favor of the version which has longer model
load duration.
By initialising timeSinceLastUpdate with 0 we ensure to not give any
variant a head start.
Storing performance.now in a variable makes it also marginally fairer
and cleaner in general.